Show market type only in closed/open markets pages
